Update README.md
Twitter sentiment analysis is performed to identify the sentiments of the people towards various topics. For this project, I will be analysing the sentiment of people towards Amazon customer service and new launched app. I will be using the data available on Kaggle to create this machine learning model. The collected tweets from Twitter will be analysed using machine learning to identify the different sentiments present in the tweets. The different sentiments identified in this project include positive sentiment, negative sentiment and neutral sentiment. I will also be using different classifiers to see which classifier gives the best model accuracy.